refactor: use correct imports in viewmodels and listeners
