+++ app/src/main/java/com/moscow/tudee/di/presentationModule.kt
@@ -1,7 +1,7 @@
 package com.moscow.tudee.di
 
-import com.moscow.tudee.presentation.category.categoryScreen.CategoryViewModel
-import com.moscow.tudee.presentation.category.categoryTasksScreen.CategoryTasksViewModel
+import com.moscow.tudee.presentation.screen.category.categoriesScreen.CategoryViewModel
+import com.moscow.tudee.presentation.screen.category.categoryTasksScreen.CategoryTasksViewModel
 import com.moscow.tudee.presentation.screen.home.HomeViewModel
 import com.moscow.tudee.presentation.task.TaskViewModel
 import com.moscow.tudee.presentation.ui.splash.viewmodel.SplashViewModel
